apps/api/requirements.txt
PyJWT>=2.8.0
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-uvicorn
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-psycopg2

opentelemetry-sdk

opentelemetry-instrumentation-starlette

opentelemetry-instrumentation-httpx

redis>=5

email-validator>=2
prometheus-fastapi-instrumentator==6.7.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-starlette==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-util-http==0.57b0
